Incremento dos logs

+++ index.js
@@ -11,7 +11,7 @@ var shell = require('shelljs');
 var notifier = require('node-notifier');
 var logger = require('electron-logger');
 
-logger.setOutput({file: "./tmp.log"});
+logger.setOutput({file: path.resolve(process.cwd(), 'updater.log')});
 
 var url = require('url');
 var https = require('https');
@@ -19,6 +19,16 @@ var https = require('https');
 const MANIFEST_DIR = path.resolve(process.cwd(), 'package.json');
 
 var manifest = require(MANIFEST_DIR);
+
+if (!manifest.autoupdater) {
+    logger.error('package.json está sem as configurações de autoupdate. Verifique');
+    throw 'package.json está sem as configurações de autoupdate. Verifique';
+}
+
+if (!manifest.autoupdater.AWS_access_key || !manifest.autoupdater.AWS_secret_acccess_key || !manifest.autoupdater.AWS_region || !manifest.autoupdater.AWS_bucket_name) {
+    logger.error('package.json está sem as configurações da AWS. Verifique');
+    throw 'package.json está sem as configurações da AWS. Verifique';
+}
 
 AWS.config.update({
     accessKeyId: manifest.autoupdater.AWS_access_key,
@@ -39,7 +49,6 @@ function isNWJS() {
     try {
         return (typeof nw !== "undefined");
     } catch (e) {
-        console.log(e);
         return false;
     }
 }
@@ -47,16 +56,21 @@ function isNWJS() {
 var app, appDir, appInstDir, appExec, bundledUpdaterPath;
 
 if (isNWJS()) {
+    logger.info('Aplicação rodando em NW.js');
     app = nw.App;
 } else {
+    logger.info('Aplicação rodando em Electron');
     app = require('electron').app;
 }
 
 shell.mkdir('-p', UPDATES_DIR);
+logger.info('Pasta temporária criada em ' + UPDATES_DIR);
 
 resolvePaths();
 
 exports.checkForUpdates = function () {
+
+    logger.info('iniciando checkUpdates');
 
     checkUpdates().then(function (contents) {
 
@@ -66,11 +80,8 @@ exports.checkForUpdates = function () {
     }).then(function (data) {
 
         if (!data.url) {
-            console.log('Sem updates!');
             return;
         }
-
-        console.log('initializing downloads: ' + data.url);
 
         var bundlePath = path.resolve(UPDATES_DIR, data.latest_version);
 
@@ -83,14 +94,12 @@ exports.checkForUpdates = function () {
             });
 
     }, function (err) {
-        console.log('Erro ao buscar updates');
-        console.log(err);
+        logger.error('Erro ao buscar updates');
+        logger.error(err);
     });
 };
 
 function checkUpdates() {
-
-    console.log('checking for updates');
 
     return new Promise(function (resolve, reject) {
 
@@ -101,19 +110,25 @@ function checkUpdates() {
             return reject("Parâmetros incompletos. Verifique");
         }
 
-        console.log('pasta: ' + platform);
+        logger.info('buscando arquivos no S3 com parâmetros');
+        logger.info('current_version: ' + current_version);
+        logger.info('platform: ' + platform);
+        logger.info('AWS_bucket_name: ' + manifest.autoupdater.AWS_bucket_name);
 
-        s3.listObjects({Bucket: manifest.autoupdater.AWS_bucket_name, Prefix: platform}).on('success', function handlePage(r) {
+        s3.listObjects({
+            Bucket: manifest.autoupdater.AWS_bucket_name,
+            Prefix: platform
+        }).on('success', function handlePage(r) {
 
             if (r.hasNextPage()) {
                 r.nextPage().on('success', handlePage).send();
             } else {
-                console.log('buscou update');
-                console.log(r.data.Contents);
+                logger.info('buscou arquivos');
                 resolve(r.data.Contents);
             }
         }).on('error', function (r) {
-            console.log(r);
+            logger.error('erro ao buscar arquivos no S3');
+            logger.error(r);
             reject(r.message);
         }).send();
     });
@@ -121,74 +136,94 @@ function checkUpdates() {
 
 function getLatest(current_version, files) {
 
-    console.log('buscando a ultima versao');
+    logger.info('buscando a ultima versao');
 
     return new Promise(function (resolve, reject) {
 
-        var ordered_files = [];
+        var result = 0, urlParams = '';
 
-        for (var i = 0; i < files.length; i++) {
-            var file = files[i];
-            if (file.Size > 0) {
-                ordered_files.push(file);
+        try {
+
+            var ordered_files = [];
+
+            for (var i = 0; i < files.length; i++) {
+                var file = files[i];
+                if (file.Size > 0) {
+                    ordered_files.push(file);
+                }
             }
+
+            logger.info('ordenando arquivos');
+
+            ordered_files.sort(function (fileA, fileB) {
+
+                logger.info('comparando: ' + fileA.Key + ' e ' + fileB.Key);
+
+                var filenameA = fileA.Key;
+                var blocksA = filenameA.split('_v');
+
+                if (!blocksA[1]) return 0;
+
+                var versionA = blocksA[1].replace('.zip', '');
+
+                var filenameB = fileB.Key;
+                var blocksB = filenameB.split('_v');
+
+                if (!blocksB[1]) return 0;
+
+                var versionB = blocksB[1].replace('.zip', '');
+
+                return versionCompare(versionB, versionA);
+            });
+
+            var latest_file = ordered_files[0];
+            var latest_file_name = latest_file.Key;
+
+            logger.info('Último arquivo processado: ' + latest_file_name);
+
+            var blocks = latest_file_name.split('_v');
+            var latest_version = blocks[1].replace('.zip', '');
+
+            urlParams = {Bucket: manifest.autoupdater.AWS_bucket_name, Key: latest_file.Key};
+            result = versionCompare(latest_version, current_version);
+
+        } catch (e) {
+            logger.error('Erro ao processar os arquivos do S3');
+            logger.error(e);
+            resolve({});
         }
-
-        console.log('ordenando');
-
-        ordered_files.sort(function (fileA, fileB) {
-
-            console.log('comparando: ' + fileA.Key + ' e ' + fileB.Key);
-
-            var filenameA = fileA.Key;
-            var blocksA = filenameA.split('_v');
-
-            if (!blocksA[1]) return 0;
-
-            var versionA = blocksA[1].replace('.zip', '');
-
-            var filenameB = fileB.Key;
-            var blocksB = filenameB.split('_v');
-
-            if (!blocksB[1]) return 0;
-
-            var versionB = blocksB[1].replace('.zip', '');
-
-            return versionCompare(versionB, versionA);
-        });
-
-        var latest_file = ordered_files[0];
-        var latest_file_name = latest_file.Key;
-        var blocks = latest_file_name.split('_v');
-        var latest_version = blocks[1].replace('.zip', '');
-
-        var urlParams = {Bucket: manifest.autoupdater.AWS_bucket_name, Key: latest_file.Key};
-
-        var result = versionCompare(latest_version, current_version);
 
         if (result > 0) {
 
             s3.getSignedUrl('getObject', urlParams, function (err, url) {
 
                 if (err) {
-                    console.log(err);
+                    logger.error('Erro ao buscar a url do arquivo de atualização');
+                    logger.error(err.message);
                     return reject(err.message);
                 }
 
+                var  nome_arquivo = APP_NAME_LOWER + '_v' + latest_version + '.zip';
+
+                logger.info('URL do arquivo de atualização: ' + url);
+                logger.info('nome do arquivo de atualização: ' + nome_arquivo);
+
                 resolve({
                     url: url,
-                    latest_version: APP_NAME_LOWER + '_v' + latest_version + '.zip'
+                    latest_version: nome_arquivo
                 });
             });
 
         } else {
-            console.log('nao tem mais nova');
+            logger.info('nao tem versão mais atualizada');
             resolve({});
         }
     });
 }
 
 function versionCompare(v1, v2, options) {
+
+    logger.info('comparando versões: ' + v1 + ' e ' + v2);
 
     var lexicographical = options && options.lexicographical,
         zeroExtend = options && options.zeroExtend,
@@ -238,7 +273,7 @@ function versionCompare(v1, v2, options) {
 
 function resolvePaths() {
 
-    console.log('resolving paths');
+    logger.info('resolvendo rotas...');
 
     if (process.platform === 'darwin') {
         appDir = path.resolve(process.execPath, '../../../../../../../');
@@ -261,15 +296,25 @@ function resolvePaths() {
             bundledUpdaterPath = path.resolve(appDir, 'resources', 'app', 'node_modules', 'autoupdater', 'updater.exe');
         }
     } else {
-
+        logger.error('Sistema não reconhecido pra habilitar o autoupdater');
+        throw 'Sistema não reconhecido pra habilitar o autoupdater';
     }
+
+    logger.info('rotas resolvidas');
+    logger.info('appDir: ' + appDir);
+    logger.info('appInstDir: ' + appInstDir);
+    logger.info('appExec: ' + appExec);
+    logger.info('bundledUpdaterPath: ' + bundledUpdaterPath);
 }
 
 function startUpdate(bundlePath) {
+
+    logger.info('copiando o updater de ' + bundledUpdaterPath + ' para ' + UPDATER_BIN);
+
     shell.cp(bundledUpdaterPath, UPDATER_BIN);
     shell.chmod(755 & ~process.umask(), UPDATER_BIN);
 
-    console.log('tentando chamar: ' + UPDATER_BIN);
+    logger.info('chamando o updater em ' + UPDATER_BIN);
 
     child_process.spawn(UPDATER_BIN, [
         '--bundle', bundlePath,
@@ -281,11 +326,14 @@ function startUpdate(bundlePath) {
         stdio: 'ignore'
     }).unref();
 
+    logger.info('saindo da aplicação e tentando chamar o executavel ' + appInstDir);
+
     app.quit();
 }
 
 function fetchUpdate(url, dest) {
     return fileExists(dest).then(function (exists) {
+        logger.info('arquivo já existe? ' + (exists ? 'sim' : 'não'));
         if (exists) {
             return Promise.resolve(dest);
         }
@@ -296,6 +344,8 @@ function fetchUpdate(url, dest) {
 function downloadFile(source, dest) {
 
     return new Promise(function (resolve) {
+
+        logger.info('Iniciando download do arquivo como host ' + url.parse(source).host + ' e path ' + url.parse(source).pathname);
 
         var options = {
             host: url.parse(source).host,
@@ -310,7 +360,7 @@ function downloadFile(source, dest) {
                 file.write(data);
                 progress(100 - (((fsize - file.bytesWritten) / fsize) * 100));
             }).on('end', function () {
-                console.log('terminou');
+                logger.info('terminou o download do arquivo');
                 file.end();
                 resolve();
             });
@@ -318,7 +368,7 @@ function downloadFile(source, dest) {
     });
 }
 function progress(percent) {
-    console.log("Download: " + parseInt(percent) + " %");
+    logger.info("andamento do download: " + parseInt(percent) + " %");
     // var progressBarWidth = percent * $element.width() / 100;
     // $element.find('div').css( "width", progressBarWidth );
 }
@@ -326,6 +376,8 @@ function progress(percent) {
 function notifyUser() {
 
     return new Promise(function (resolve) {
+
+        logger.info('notificando o usuário');
 
         var notification_body = manifest.autoupdater.notification_message_body ? manifest.autoupdater.notification_message_body : 'Clique aqui para instalar';
         var notification_title = manifest.autoupdater.notification_message_title ? manifest.autoupdater.notification_message_title : 'Uma nova versão está disponível!';
@@ -344,10 +396,12 @@ function notifyUser() {
         notifier.notify(options);
 
         notifier.on('click', function (notifierObject, options) {
+            logger.info('usuário clicou em sim, atualizando...');
             resolve(true);
         });
 
         notifier.on('timeout', function (notifierObject, options) {
+            logger.info('sem resposta se é pra atualizar ou não');
             resolve(false);
         });
     });
@@ -355,8 +409,11 @@ function notifyUser() {
 
 function removeFile(filepath) {
     return new Promise(function (resolve, reject) {
+        logger.info('removendo arquivo ' + filepath);
         fs.unlink(filepath, function (err) {
             if (err) {
+                logger.error('Erro ao remover arquivo');
+                logger.error(err);
                 return reject(err);
             }
             resolve();
@@ -366,14 +423,19 @@ function removeFile(filepath) {
 
 function fileExists(bundledUpdaterPath) {
     return new Promise(function (resolve) {
+        logger.info('verificando se arquivo existe em ' + bundledUpdaterPath);
         fs.stat(bundledUpdaterPath, function (err, stats) {
             if (err) {
+                logger.error('Erro ao verificar se arquivo existe');
+                logger.error(err);
                 if (err.code === 'ENOENT') {
                     return resolve(false);
                 }
                 throw err;
             }
-            if (stats.size < 5000) {
+
+            if (stats.isFile() && stats.size < 5000) {
+                logger.info('arquivo existente mas com tamanho inválido: ' + stats.size + ', removendo...');
                 return removeFile(bundledUpdaterPath).then(function () {
                     return resolve(false);
                 });
